requirements.txt
python-telegram-bot==20.7
openai==1.12.0
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
tiktoken==0.5.2
yoomoney==0.1.2

